requirements.txt
pandas>=2.0
numpy>=1.24
nba_api>=1.4
scikit-learn>=1.3
fastapi>=0.104
uvicorn>=0.24
anthropic>=0.18
python-dotenv>=1.0
apscheduler>=3.10
psycopg2-binary>=2.9
joblib>=1.3
requests>=2.31
tqdm>=4.66
sqlalchemy>=2.0
httpx>=0.25
